Added Direct_Write and adding editor stuff like Window, Button, List_Box ...

+++ src/win32/win_types.h
@@ -3,15 +3,16 @@
 
 #include <windows.h>
 
-typedef BYTE    u8;
-typedef WORD    u16;
-typedef UINT    u32;
-typedef DWORD   u64;
+typedef UINT8    u8;
+typedef UINT16   u16;
+typedef UINT32   u32;
+typedef UINT64   u64;
 
 typedef INT8    s8;
 typedef INT16   s16;
 typedef INT32   s32;
 typedef INT64   s64;
+
 
 typedef float  float32;
 typedef double float64;
